packages/ui/src/db/adapters/TableAdapter.test.ts
import { TableAdapter } from './TableAdapter';

describe('section/adapters/TableAdapter', () => {
  const table = new TableAdapter();

  it('should perform/mimic add, update & delete operations', async () => {
    let result = await table.add({ name: 'Item 1' });
    expect(result).toBe(0);
    result = await table.add({ name: 'Item 1' }, 1);
    expect(result).toBe(1);

    result = await table.update(2, { isActive: true });
    expect(table.toArray()).toEqual([{ id: 1, name: 'Item 1' }]);
    expect(result).toBe(0);
    result = await table.update(1, { isActive: true });
    expect(table.toArray()).toEqual([{ id: 1, name: 'Item 1', isActive: true }]);
    expect(result).toBe(1);

    result = await table.delete(2);
    expect(table.toArray()).toEqual([{ id: 1, name: 'Item 1', isActive: true }]);
    expect(result).toBe(0);
    result = await table.delete(1);
    expect(table.toArray()).toEqual([]);
    expect(result).toBe(1);
  });

  it('should retrieve data correctly', async () => {
    await table.add({ name: 'Item 1' }, 1);
    await table.add({ name: 'Item 2' }, 2);

    let result: unknown | unknown[] = table.where({ name: 'Item 1' }).toArray();
    expect(result).toEqual([{ name: 'Item 1', id: 1 }]);

    result = await table.get(1);
    expect(result).toEqual({ name: 'Item 1', id: 1 });

    const count = await table.count();
    expect(count).toBe(2);
  });
});
